# simplify quoteString (make its code more uniform)