Benchmark results comparison:

```
benchmark                              old ns/op     new ns/op     delta
BenchmarkValidIdentifier-4             250           241           -3.60%
BenchmarkValidIdentifierUnquoted-4     423           418           -1.18%
BenchmarkValidIdentifierReserved-4     22.0          26.9          +22.27%
BenchmarkBigJSON-4                     84351117      83896179      -0.54%
BenchmarkStatementsLess-4              91.9          90.7          -1.31%
BenchmarkFill-4                        10096         9877          -2.17%
BenchmarkValueTokenFromInterface-4     2274          1977          -13.06%
```

(the speedup seems to be mostly thanks to the Fprintf)