@akavel 2019-04-03
Hi @tomnomnom ! You're welcome, happy you like it, and no problem at all; not sure if that'll make you happier or sadder, but actually I totally forgot about it since ;P Cheers and good luck, and best wishes for you to stay in good mental health! I know where you're coming from, I think burnout is kinda semi-officially becoming an occupational disease in our path of work, or something... :/